Refactor DJS-47 chain api refactor; usages update
