+++ src/modules/proposal.ts
@@ -8,10 +8,11 @@ import {
 import {Memo, Operations} from '../model/transaction';
 import {Transaction} from '../transaction';
 import {Asset} from '../model/account';
-import {ChainApi, ChainMethods} from '../api/chain';
+import {ChainApi} from '../api/chain';
 import {Utils} from '../utils';
 import {CryptoUtils} from '../crypt';
 import {ApiConnector} from '../api/apiConnector';
+import {ChainMethods} from '../api/model/chain';
 
 export class ProposalModule extends ApiModule {
     private _chainApi: ChainApi;
@@ -40,12 +41,12 @@ export class ProposalModule extends ApiModule {
         proposerAccountId: string, fromAccountId: string, toAccountId: string, amount: number, assetId: string, memoKey: string,
         expiration: string, privateKey: string): Promise<any> {
         return new Promise<any>(((resolve, reject) => {
-            const operations = new ChainMethods();
-            operations.add(ChainMethods.getAccount, fromAccountId);
-            operations.add(ChainMethods.getAccount, toAccountId);
-            operations.add(ChainMethods.getAsset, assetId);
-
-            this._chainApi.fetch(operations)
+            const operations = [].concat(
+                new ChainMethods.GetAccount(fromAccountId),
+                new ChainMethods.GetAccount(toAccountId),
+                new ChainMethods.GetAsset(assetId)
+            );
+            this._chainApi.fetch(...operations)
             .then(result => {
                 const [senderAccount, receiverAccount, asset] = result;
                 const senderAccountObject = JSON.parse(JSON.stringify(senderAccount));
